Creating TaskGatewayInterface
Because we are not really sure what kind of implementations we're going to need, we use an interface to define the methods we want to use without actually implementing them.

We update our `App\Service\TaskService` and `App\Test\Service\TaskServiceTest` to ensure they are using the interface for stubbing.